+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 ## 30.0.10
 New:
-- `GeoCalendar`: Added 'isToDateOnly' and 'isFromDateOnly' props to force to select only one side of the range
-- `GeoCalendarDropdown`: Added 'isToDateOnly' and 'isFromDateOnly' props to force to select only one side of the range
+- `GeoCalendar`: Added 'isToDateDisabled' and 'isFromDateDisabled' props to force to select only one side of the range
+- `GeoCalendarDropdown`: Added 'isToDateDisabled' and 'isFromDateDisabled' props to force to select only one side of the range
 ## 30.0.9
 Fix:
 - `GeoTable`: Remove unnecessary height: 100% attribute in table-row-headers styles that could break header border lines in some situations